Milestone_3/requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langgraph
faiss-cpu
python-dotenv
